chore: Update stale.yml to only apply to pulls, not issues (#200)
Since we're using GH issues to track work (thanks @misskniss and @tjcran for setting this up!) - we probably don't want the stale bot to close untouched issues every month 😬 

I considered using a separate stale-duration (like, 365 days for issues) - but just turned it off now. Easy for us to change if we want to adjust the process.