feat(site): add UserCell component
Summary:

This is a first step in porting over v1 AuditLog in a refactored/cleaned
up fashion. This isn't a direct port, since we do not yet have a
UserAvatar component.

Details:

- Port over UserCell from v1, sans UserAvatar impl
- Add tests and stories for UserCell

Notes:

We do not have a holistic solution for handling localization, but
starting from some kind of easy way that collects/resources strings will
make the migration significantly easier. It will also help out our
product copy owner, @khorne3 with maintenance. An RFC regarding this
might be necessitated.

Impact:

This change does not have any user-facing impact yet, because the
UserCell is not yet rendered in the product. This enables an incremental
approach to migrating in the FE of the Audit Log, which is still waiting
on the BE port.

Relations:

- This commit relates to #472, but does not finish it.
- This commit should not merge until after #465 and #483 because it's
based on them.