feat(site): add ActionCell component
Summary:

This is a direct follow-up to #484 and a next step in porting over the
AuditLog with appropriate refactoring. This isn't a direct port; we used
to have a map of icons on the FE, but we want to no loner split the
logic between FE and BE. If we want icons at a later time point, we can
return an icon identifier in the response, or we can simplify the view
and gain real estate by omitting the icons. At the time of this commit,
I prefer omitting the icons.

Details:

- Port over ActionCell from v1, sans icons
- Add tests and stories for ActionCell

Impact:

This change does not have any user-facing impact yet, because the
ActionCell is not yet rendered in the product. This enables an incremental
approach to migrating in the FE of the Audit Log, which is still waiting
on the BE port.

Relations:

- This commit relates to #472, but does not finish it.
- This commit should not merged until after #484 because it's based from
it