feat(site): add TargetCell component
Summary:

This is a direct follow-up to #484 and #500. It is a part of many steps
in porting/refactoring the AuditLog from v1.

Details:

- Port over TargetCell from v1, with refactorings
- Add tests and stories

Impact:

This change does not have any user-facing impact yet because AuditLog is
not yet available in the product. This is part of an incremental
approach; the FE is still waiting on the BE port.

Relations:

- This commit relates to #472, but does not finish it
- This commit should not be merged until after #484 and #500, because it
builds off of them.