Merge "StatusCell" with "ActionCell"
The old AuditLog had bad use of screen real estate. This change merges
the old status cell (that had a large chip) with the action cell. These
two cells seem related (wanting to see the action and whether or not it
was successful in one spot).

NTS: squash this message in before merge