feat(site): add TimeCell component
Summary:

This is direct follow-up to #484, #500, and #501. It is a part of many
steps in porting/refactoring the AuditLog from v1.

Details:

- Port over TimeCell from v1, with refactorings
- A jest test was not added because we use locale dates and times, which
can be tricky to test. This can be addressed in the future.

Impact:

This change does not have any user-facing impact yet because AuditLog is
not yet available in the product. This is part of an incremental
approach; the FE is still waiting on the BE port.

Relations:

- This commit relates to #472, but does not finish it
- This commit should not be merged until after #484, #500 and #501,
because it builds off of them