refactor: wrap codersdk client returned by coderdtest.New()
Currently, coderdtest.New() returns the bare codersdk.Client
to talk to the in-memory server and nothing else. This means
there is currently no way to pass additional internal 'knobs'
to unit tests.

This commit creates a new struct coderdtest.API and changes
the return type of coderdtest.New() to return this type.
updates references to the client. Embedding the client was
considered and rejected in favour of just keeping it as a
struct field for clarity.

Currently no additional fields are exposed by coderdtest.API
but it is now trivial to add more fields.