fix: Decrease postgres test timeout (`make test-postgres`)
This commit lowers the postgres test timeout from 30m to 20m, currently
our postgres tests seem to take 8-10m, a 2x factor should suffice.
Comments were updated in both places to reflect the reasoning and
necessity of keeping these values in sync.

They used to take longer but the `count` was lowered in
3d40cb85b7e875cd8677d5c857db25351b514a84.

The actual timeout value of `make test-postgres` got overlooked in
https://github.com/coder/coder/pull/3079.