Open up frontend PR reviews
It's more difficult to get changes into the frontend on off hours (e.g weekend) due to the requirement. Since the backend lacks a symmetrical requirement, I figured it would be OK to lift the requirement for frontend too. As an added benefit, I think lifting the gate steps us closer to a more cross functional team.