fix: Allow terraform provisions to be gracefully cancelled
This change allows terraform commands to be gracefully cancelled on
Unix-like platforms by signaling interrupt on provision cancellation.

One implementation detail to note is that we do not necessarily kill a
running terraform command immediately even if the stream is closed. The
reason for this is to allow for graceful cancellation even in such an
event. Currently the timeout is set to one minute, which was chosen
arbitrarily.

Also note that the `force` flag was added to `provisioner.proto` and is
handled in the `terraform` package, however, it is not used by the
`provisionerd/runner`. The reason is that the `runner` would need to be
refactored. Currently the "force stop" context is used as a base for
streams, and this would mean that in a force stop event, the stream
would close and we can't send out our force stop cancellation message.

Related: #2683

The above issue may be partially of fully fixed by this change.