fix: Start SFTP sessions in user home (working directory)
This is a simple fix for #3620 by changing the working directory of
`coder agent` at runtime.

Since the agent doesn't write files to it's working directory, I can't
think of any downsides to this method. It does not, however, exclude
that we rewrite the SFTP implementation to use the modernized
`RequestServer` sometime in the future.

Opted to place this in agent init vs `cli/agent` because of portability
of fix, tests don't usually use `cli/agent`.

Fixes #3620