fix: agent stats websocket blocking until next interval
For `httpapi.Heartbeat` to work properly, a call to `wsjson.Read` must
always be in flight, otherwise pings can't be exchanged. This caused
the ping loop to hang until the next stat interval. Since the loop was
blocking until the next interval it would also never exit until it
fired.