refactor: extend Overrides for MuiSkeleton (#4818)
* refactor: extend Overrides for MuiSkeleton

Looking at the types provided with `@material-ui/core`, it is indeed
missing `MuiSkeleton`. I'm not sure why. I found this issues upstream: https://github.com/mui/material-ui/issues/24959

I also tried upgrading the package to the latest - 4.12.4 I believe but
that didn't fix it either.

I resorted to extending the `Overrides` type based on what I saw in the
most recent version of the declaration file. This is a temporary fix but
opts back in to type safety instead of resorting to `@ts-ignore`.

* formatting