fix(site): Use correct UUID for web terminal when first opened
While working on #5403 I noticed the web terminal connects with a
different UUID than what is serialized in the URL, this resulted in a
new session being opened if I refreshed the page.

The reasons seems to be that we were accessing raw DOM elements
(`location.search`) for navigation and omitting `react-router-dom`. This
resulted in a new navigation and UUID being generated after the fact.