chore: Add ability to hash an authorizationc all to a unqiue hash key.
This is useful for caching purposes.
Order matters for the hash key for slices, so a slice with
the same elements in a different order is a different hash key.
Athlough they are functionally equivalent, our slices have deterministic
orders in the database, so our subjects/objects will match.
There is no need to spend the time or memory to sort these lists.