Poll for service banner
We need to provide the banner in the test so I changed the manifest
argument to support passing in the client.

We could not set the banner func afterward like I was doing before
because by that point the banner will already have fetched and will not
update until the next interval.

The test interval func also changes since I wanted to have a 15 second
interval but that meant 1.5 seconds in the test which is way too long.