Remove error from close return
Some errors are returned and some are logged at the debug level.  The
reason for the debug logging is that the process might already be closed
so in that case I did not want to return an error.  This will be fixed
in a future commit but since this error is only logged by the caller
anyway, go ahead and just log it in the backend to simplify things.