Do not hold mutex while waiting for state in screen
Holding off on changing this for the buffered version until a potential
issue with the pty closing before we add to the map is sorted.

I moved the heartbeat start out as well since while we want to wait to
start that until after ready, it does not need to be behind the mutex.