Do not hold mutex while waiting for state
Since we might add to the map after it closes, tweak the delete to also
run when the pty closes.

I think this also fixes that issue where if you canceled the context you
pass into Attach it would not do anything since readConnLoop does not
use the context for anything but logging.