Immediately read screen process
It is possible for the process to immediately exit (for example if you
run `echo hello`), then we would wait for the `version` command to
succeed but it never will because the session is already gone.

If we immediately read to the process then we can tell when it has gone
away and we can abort.