Remove err from wait callback
We do not really do anything with it other than just return it, and if
we do need to do something with it we can just do it on the return from
`waitForStateOrContext`, we probably would not need to use it behind the
mutex.

Also we should check the state outside since there is technically no
guarantee an error is set on a state change (although in practice for
close/done there always is, maybe this should be enforced in some way).